requirements.txt
hdbscan==0.8.29
matplotlib==3.5.1
numpy==1.23.5
Pillow==9.0.0
scikit_learn==1.2.2
torch==1.12.1+cu116
torchvision==0.13.1+cu116
cupy_cuda12x==12.0.0

--extra-index-url https://pypi.nvidia.com
cudf_cu11==23.2.0
cuml_cu11==23.2.0
